Additional changes for SDSoC support
This project is still incomplete and in development.

-Changed video input pipeline to use 8-bit grayscale input. This reduces
DDR bandwidth used by the input pipeline. This will likey be changed to
YUYV (aka YUV422) input with subsampling before launch so that color
input is possible.

-Changed clock speed of AXI memory interconnect to ~120 MHZ. This allows
the design to meet timing, but makes 1080p input or output not possible
because the axi stream buses are also currently clocked with this clock.
This should probably be changed before launch to support 1080p again (at
least on the output).

-Added notes to SDSoC staging area describing that Linux support is
currently broken.

-Remove lib_bsp because you are not supposed to include libxil.a in
SDSoC 2016.4 because it will automatically generate a BSP for you. For
some reason you still need to include the "include" folder though.